+++ packages/react-contexts/src/middlewares/refresh-session.ts
@@ -1,9 +1,4 @@
-/* eslint-disable no-console */
-import {
-  ResponseCookies,
-  RequestCookies,
-  type ResponseCookie,
-} from 'next/dist/compiled/@edge-runtime/cookies'
+import { type ResponseCookie } from 'next/dist/compiled/@edge-runtime/cookies'
 import {
   NextFetchEvent,
   NextMiddleware,
@@ -19,6 +14,8 @@ import {
 } from '@titicaca/constants'
 
 import { parseApp } from '../user-agent-context'
+
+import { applySetCookie } from './utils/apply-set-cookie'
 
 export function refreshSessionMiddleware(next: NextMiddleware) {
   return async function middleware(
@@ -92,23 +89,4 @@ export function refreshSessionMiddleware(next: NextMiddleware) {
 
 function deriveAllCookies(cookies: { name: string; value: string }[]) {
   return cookies.map(({ name, value }) => [name, value].join('=')).join('; ')
-}
-
-/** Reference: https://github.com/vercel/next.js/discussions/50374#discussioncomment-6732402 */
-function applySetCookie(req: NextRequest, res: NextResponse) {
-  const setCookies = new ResponseCookies(res.headers)
-  const newReqHeaders = new Headers(req.headers)
-  const newReqCookies = new RequestCookies(newReqHeaders)
-  setCookies.getAll().forEach((cookie) => newReqCookies.set(cookie))
-
-  const dummyRes = NextResponse.next({ request: { headers: newReqHeaders } })
-
-  dummyRes.headers.forEach((value, key) => {
-    if (
-      key === 'x-middleware-override-headers' ||
-      key.startsWith('x-middleware-request-')
-    ) {
-      res.headers.set(key, value)
-    }
-  })
 }
